+++ examples/minireview.rs
@@ -7,8 +7,12 @@ use futures::{
 };
 use ipnet::IpNet;
 use review_database::{migrate_data_dir, Database, Store};
-use review_web::{self as web, graphql::AgentManager, CertManager};
-use roxy::ResourceUsage;
+use review_web::{
+    self as web,
+    graphql::AgentManager,
+    graphql::{Process, ResourceUsage},
+    CertManager,
+};
 use serde::Deserialize;
 use std::{
     collections::HashMap,
@@ -93,6 +97,10 @@ impl AgentManager for Manager {
 
     async fn send_and_recv(&self, _key: &str, _msg: &[u8]) -> Result<Vec<u8>, Error> {
         bail!("Not supported")
+    }
+
+    async fn get_process_list(&self, hostname: &str) -> Result<Vec<Process>, Error> {
+        bail!("Host {hostname} is unreachable")
     }
 
     async fn get_resource_usage(&self, hostname: &str) -> Result<ResourceUsage, Error> {